Add more tests.
Make a few changes as per @cdeepakroy's suggestions.

Refactored the tests to allow better parallel testing and less code duplication.
