+++ plugin.cmake
@@ -49,7 +49,6 @@ add_python_test(cache PLUGIN large_image BIND_SERVER)
 add_python_test(tiles PLUGIN large_image BIND_SERVER EXTERNAL_DATA
   "plugins/large_image/sample_image.ptif"
   "plugins/large_image/sample_svs_image.TCGA-DU-6399-01A-01-TS1.e8eb65de-d63e-42db-af6f-14fefbbdf7bd.svs"
-  "plugins/large_image/sample_jp2k_33003_TCGA-CV-7242-11A-01-TS1.1838afb1-9eee-4a70-9ae3-50e3ab45e242.svs"
   )
 set_property(TEST server_large_image.tiles APPEND PROPERTY ENVIRONMENT
   "LARGE_IMAGE_DATA=${PROJECT_BINARY_DIR}/data/plugins/large_image")
@@ -68,6 +67,14 @@ set_property(TEST server_large_image.cached_tiles.PythonCache APPEND PROPERTY
   ENVIRONMENT
   "LARGE_IMAGE_CACHE_BACKEND=python"
   "LARGE_IMAGE_DATA=${PROJECT_BINARY_DIR}/data/plugins/large_image")
+
+add_python_test(sources PLUGIN large_image BIND_SERVER EXTERNAL_DATA
+  "plugins/large_image/sample_image.ptif"
+  "plugins/large_image/sample_jp2k_33003_TCGA-CV-7242-11A-01-TS1.1838afb1-9eee-4a70-9ae3-50e3ab45e242.svs"
+  )
+set_property(TEST server_large_image.sources APPEND PROPERTY ENVIRONMENT
+  "LARGE_IMAGE_DATA=${PROJECT_BINARY_DIR}/data/plugins/large_image")
+
 
 add_python_test(import PLUGIN large_image BIND_SERVER)
